Increase max. number of labels in data-flow sanitizer
Use 32-bits unsigned values instead of 16 bits to represent data-flow sanitizer
labels, and increase the shadow memory region for Linux/x86_64 accordingly. Set
the max. number to a million labels instead of 2^16. This is needed when using
'-dfsan-discovery' to allow non-trivial program executions.

+++ compiler-rt/lib/dfsan/dfsan.h
@@ -19,10 +19,10 @@
 #include "dfsan_platform.h"
 
 using __sanitizer::uptr;
-using __sanitizer::u16;
+using __sanitizer::u32;
 
 // Copy declarations from public sanitizer/dfsan_interface.h header here.
-typedef u16 dfsan_label;
+typedef u32 dfsan_label;
 
 struct dfsan_label_info {
   dfsan_label l1;
